Moved all the pipeline functions into the /scripts folder. Now the scr/ folder only store functions
