+++ scripts/run_data_aquisition.py
@@ -5,14 +5,8 @@ sys.path.insert(0,'.')
 
 from src.data.make_dataset import *
 
-with open("src/configuration/logging_config.yaml", 'r') as f:  
-    logging_config = yaml.safe_load(f.read())
-    logging.config.dictConfig(logging_config)
-    logger = logging.getLogger(__name__)
-
-with open("src/configuration/project_config.yaml", 'r') as f:  
-    config = yaml.safe_load(f.read())
-
-logger.info("Downloading the raw dataset...")
-stock_df = make_dataset(ticker_list, PERIOD, INTERVAL)
-logger.info("Finished downloading the raw dataset!")
+if __name__ == "__main__":
+    
+    logger.info("Downloading the raw dataset...")
+    stock_df = make_dataset(ticker_list, PERIOD, INTERVAL)
+    logger.info("Finished downloading the raw dataset!")
