--- scripts/test.py
@@ -1,6 +1,0 @@
-print("HELLO WORRRRRRDDDDDDDDDDDDDddd")
-print("HELLO WORRRRRRDDDDDDDDDDDDDddd")
-print("HELLO WORRRRRRDDDDDDDDDDDDDddd")
-print("HELLO WORRRRRRDDDDDDDDDDDDDddd")
-print("HELLO WORRRRRRDDDDDDDDDDDDDddd")
-print("HELLO WORRRRRRDDDDDDDDDDDDDddd")
