+++ src/models/evaluate_model.py
@@ -79,59 +79,6 @@ def evaluate_and_store_performance(model_type, ticker, y_true, y_pred, latest_pr
         results_df.to_csv(file_path, index=False)
 
 
-def daily_model_evaluation(model_type=None, ticker=None):
-    """Performs the models' performance daily evaluation"""
-
-    # Loads the out of sample forecast table and the training dataset
-    current_train_df = pd.read_csv(os.path.join(PROCESSED_DATA_PATH, 'processed_df.csv'), parse_dates=["DATE"])
-    historical_forecasts_df = pd.read_csv(os.path.join(OUTPUT_DATA_PATH, 'forecast_output_df.csv'), parse_dates=["DATE","RUN_DATE"])
-
-    # TARGET_COL = config["model_config"]["TARGET_COL"]
-    # PREDICTED_COL = config["model_config"]["PREDICTED_COL"]
-    available_models = model_config['available_models']
-
-    latest_price_date = current_train_df["DATE"].max().date()
-    latest_run_date = historical_forecasts_df["RUN_DATE"].max().date()
-
-    logger.debug(f"Latest availabe date: {latest_price_date}")
-    logger.debug(f"Latest run date: {latest_run_date}")
-
-    # Check the ticker parameter
-    if ticker:
-        ticker = ticker.upper() + '.SA'
-        historical_forecasts_df = historical_forecasts_df[historical_forecasts_df[CATEGORY_COL] == ticker]
-        current_train_df = current_train_df[current_train_df[CATEGORY_COL] == ticker]
-    
-    # Check the model_type parameter 
-    if model_type is not None and model_type not in available_models:
-        raise ValueError(f"Invalid model_type: {model_type}. Choose from: {available_models}")
-    
-    elif model_type:
-        available_models = [model_type]
-
-    for ticker in historical_forecasts_df[CATEGORY_COL].unique():
-        ticker_hist_forecast = historical_forecasts_df[historical_forecasts_df[CATEGORY_COL] == ticker]
-        ticker_train_df = current_train_df[current_train_df[CATEGORY_COL] == ticker]
-
-        for model_type in available_models:
-            model_type = model_type.upper()
-
-            latest_value = ticker_train_df[TARGET_COL].values[-1]
-
-            try:
-                predicted_value = ticker_hist_forecast[
-                    (ticker_hist_forecast["MODEL_TYPE"] == model_type) \
-                    & (ticker_hist_forecast["RUN_DATE"] == pd.to_datetime(latest_run_date)) \
-                    & (ticker_hist_forecast["DATE"] == pd.to_datetime(latest_price_date))
-                ][PREDICTED_COL].values[0]
-
-            except:
-                logger.error(f"\nThe last forecasts where made at {latest_run_date}, A.K.A today. Comeback tomorrow in order to calculate today's performance.")
-                raise ValueError(f"\nThe last forecasts where made at {latest_run_date}, A.K.A today. Comeback tomorrow in order to calculate today's performance.")
-
-            evaluate_and_store_performance(model_type, ticker, latest_value, predicted_value, latest_price_date, latest_run_date)
-
-
 def update_test_values(X: pd.DataFrame, y: pd.Series, day: int) -> tuple[pd.DataFrame, pd.Series]:
     """
     Prepares the feature and target data for testing on a specific day.
@@ -270,75 +217,3 @@ def stepwise_prediction(X: pd.DataFrame, y: pd.Series, forecast_horizon: int, mo
     #learning_curves_fig, feat_imp = extract_learning_curves(best_model, display=True)
     
     return pred_df, X_testing_df
-
-
-def model_crossval_pipeline(tune_params, model_type, ticker):
-
-    # TARGET_COL = config["model_config"]["TARGET_COL"]
-    # global PREDICTED_COL
-    # PREDICTED_COL = config["model_config"]["PREDICTED_COL"]
-    # FORECAST_HORIZON = config['model_config']['forecast_horizon']
-    # available_models = config['model_config']['available_models']
-
-    validation_report_df = pd.DataFrame()
-
-    logger.debug("Loading the featurized dataset..")
-    feature_df = pd.read_csv(os.path.join(PROCESSED_DATA_PATH, PROCESSED_DATA_NAME), parse_dates=["DATE"])
-
-    # Check the ticker parameter
-    if ticker:
-        ticker = ticker.upper() + '.SA'
-        feature_df = feature_df[feature_df[CATEGORY_COL] == ticker]
-        
-    # Check the model_type parameter 
-    if model_type is not None and model_type not in available_models:
-        raise ValueError(f"Invalid model_type: {model_type}. Choose from: {available_models}")
-    
-    elif model_type:
-        available_models = [model_type.upper()]
-
-    for ticker in feature_df[CATEGORY_COL].unique():
-        filtered_feature_df = feature_df[feature_df[CATEGORY_COL] == ticker].copy().drop(CATEGORY_COL, axis=1)
-        
-        for model_type in available_models:
-            logger.info(f"Performing model cross validation for ticker symbol [{ticker}] using model [{model_type}]...")
- 
-            predictions_df, X_testing_df = stepwise_prediction(
-                X=filtered_feature_df.drop(columns=[TARGET_COL], axis=1),
-                y=filtered_feature_df[TARGET_COL],
-                forecast_horizon=FORECAST_HORIZON,
-                model_type=model_type,
-                ticker=ticker,
-                tune_params=tune_params
-            )
-
-            predictions_df[CATEGORY_COL] = ticker
-            predictions_df["TRAINING_DATE"] = dt.datetime.today().date()
-            validation_report_df = pd.concat([validation_report_df, predictions_df], axis=0)
-    
-    logger.info("Writing the testing results dataframe...")
-    validation_report_df.to_csv(os.path.join(OUTPUT_DATA_PATH, CROSS_VAL_DATA_NAME), index=False)
-
-
-if __name__ == "__main__":
-
-    parser = argparse.ArgumentParser(description="Perform Out-of-Sample Tree-based models Inference.")
-
-    parser.add_argument(
-        "-mt", "--model_type",
-        type=str,
-        choices=["XGB", "ET"],
-        help="Model name use for inference (XGB, ET) (optional, defaults to all)."
-    )
-    parser.add_argument(
-        "-ts", "--ticker",
-        type=str,
-        help="""Ticker Symbol for inference. (optional, defaults to all).
-        Example: bova11 -> BOVA11.SA | petr4 -> PETR4.SA"""
-    )
-    args = parser.parse_args()
-
-    logger.info("Starting the Daily Model Evaluation pipeline...")
-    daily_model_evaluation(args.model_type, args.ticker)
-    # model_crossval_pipeline(False, args.model_type, args.ticker)
-    logger.info("Daily Model Evaluation Pipeline completed successfully!")
